fix(31-08): CreatureObject attributesOnSet callback const size_t -> unsigned int (BITS-03)
Rule-3 blocking fix exposed once the AutoDeltaVector header became x64-compilable:
CreatureObject.cpp:608 registers attributesOnSet as an AutoDeltaVector<Attributes::
Value,CreatureObject>::setOnSet callback, but the member fn took (const size_t elem,
...). size_t == unsigned int only on 32-bit; on x64 size_t (8B) != unsigned int (4B),
so the pointer-to-member type mismatched setOnSet's deliberately wire-stable
(const unsigned int, ...) contract -> C2664. Retype the elem param to unsigned int
in both the header decl and the .cpp def (elem is only used in an == comparison;
no width assumption). Mirrors the 31-07 GroupObject members/memberShips callback fix.
Verified: CreatureObject.cpp compiles x64 exit 0.

+++ src/engine/client/library/clientGame/src/shared/object/CreatureObject.cpp
@@ -2923,7 +2923,7 @@ void CreatureObject::skillModsOnErase         (const std::string &, const std::p
 
 //----------------------------------------------------------------------
 
-void CreatureObject::attributesOnSet (const size_t elem, const Attributes::Value & oldValue, const Attributes::Value & newValue)
+void CreatureObject::attributesOnSet (const unsigned int elem, const Attributes::Value & oldValue, const Attributes::Value & newValue)
 {
 	if (newValue > oldValue)
 		return;
